refactor: move resolveReplyPreview out of the page module (#4245)
CI lint ratchet caught react-refresh/only-export-components: a module
whose default export is a component must not also export non-components,
or Fast Refresh breaks. Exporting the helper from UnifiedMessagesPage.tsx
purely to make it testable was the wrong seam.

Moved to src/utils/replyPreview.ts and made generic over the message
type, so it carries no coupling to the page's UnifiedMessage shape. The
test moves with it and no longer needs the jsdom environment, since it
no longer imports the page (and its CSS) just to reach a pure function.

No behavior change.
